feat(core/pipeline): Create a preconfiguredJob.reader
