+++ app/scripts/modules/core/src/pipeline/config/stages/preconfiguredJob/PreconfiguredJobStageConfig.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { set } from 'lodash';
 
 import { IStageConfigProps, StageConfigField } from 'core/pipeline';
-import { IPreconfiguredJobParameter } from './preconfiguredJobStage';
+import { IPreconfiguredJobParameter } from './preconfiguredJob.reader';
 
 export class PreconfiguredJobStageConfig extends React.Component<IStageConfigProps> {
   private parameterFieldChanged = (fieldIndex: string, value: any) => {
