+++ app/scripts/modules/core/src/pipeline/config/stages/preconfiguredJob/preconfiguredJobStage.ts
@@ -1,47 +1,21 @@
 import { module } from 'angular';
-
-import { API } from 'core/api/ApiService';
+import { fromPairs } from 'lodash';
 import { Registry } from 'core/registry';
-
 import { ExecutionDetailsTasks } from '../common';
 import { PreconfiguredJobExecutionDetails } from './PreconfiguredJobExecutionDetails';
 import { PreconfiguredJobStageConfig } from './PreconfiguredJobStageConfig';
-
-export interface IPreconfiguredJobParameter {
-  name: string;
-  label: string;
-  description?: string;
-  type: string;
-  defaultValue?: string;
-}
-
-interface IPreconfiguredJob {
-  type: string;
-  label: string;
-  noUserConfigurableFields: boolean;
-  description?: string;
-  waitForCompletion?: boolean;
-  parameters?: IPreconfiguredJobParameter[];
-  producesArtifacts: boolean;
-}
+import { PreconfiguredJobReader } from './preconfiguredJob.reader';
 
 export const PRECONFIGUREDJOB_STAGE = 'spinnaker.core.pipeline.stage.preconfiguredJobStage';
 
 module(PRECONFIGUREDJOB_STAGE, []).run(() => {
-  API.one('jobs')
-    .all('preconfigured')
-    .getList()
-    .then((preconfiguredJobs: IPreconfiguredJob[]) => {
-      preconfiguredJobs.forEach(preconfiguredJob => {
+  PreconfiguredJobReader.list().then(preconfiguredJobs => {
+    preconfiguredJobs
+      .filter(job => job.uiType !== 'CUSTOM')
+      .forEach(preconfiguredJob => {
         const { label, description, type, waitForCompletion, parameters, producesArtifacts } = preconfiguredJob;
-        const defaults = {
-          parameters: parameters.reduce((acc, parameter) => {
-            if (parameter.defaultValue) {
-              acc[parameter.name] = parameter.defaultValue;
-            }
-            return acc;
-          }, {} as any),
-        };
+        const defaults = fromPairs(parameters.filter(p => p.defaultValue).map(p => [p, name, p.defaultValue]));
+
         Registry.pipeline.registerStage({
           label,
           description,
@@ -59,5 +33,5 @@ module(PRECONFIGUREDJOB_STAGE, []).run(() => {
           producesArtifacts,
         });
       });
-    });
+  });
 });
